REFACTOR: move admin actions to discrete component.
